Add 30-minute snap interval option for Oracle AWR generation
settings.html: added 'Every 30 minutes' (value=0.5) to interval dropdown
awr_oracle_connections.sql: snap_interval_hrs changed INTEGER -> NUMERIC(4,1)
oracle_awr_fetcher.py: snap_interval_hrs parsed as float() not int()
Table display: shows 'Every 30 min' for 0.5 value

+++ modules/oracle_awr_fetcher.py
@@ -129,7 +129,7 @@ def save_connection(data: dict, added_by: str = 'admin') -> dict:
     service_name      = (data.get('service_name') or '').strip()
     username          = (data.get('username') or '').strip()
     password          = (data.get('password') or '').strip()
-    snap_interval_hrs = int(data.get('snap_interval_hrs') or 1)
+    snap_interval_hrs = float(data.get('snap_interval_hrs') or 1)
     enabled           = bool(data.get('enabled', True))
 
     if not db_name:
